+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Report on partial disk writes when streaming envelopes to file, which previously left truncated envelopes on disk and reported success. ([#1804](https://github.com/getsentry/sentry-native/pull/1804))
 
+**Internal**:
+
+- Refactor envelope writers to better support failure tracking on each layer and not push check responsibility to client code.  ([#1807](https://github.com/getsentry/sentry-native/pull/1807))
+
 ## 0.15.0
 
 **Breaking**: